✨ feat: Add Postman collections for Quality of Software (S206)
This commit adds two new Postman collection files. These collections contain API tests for the Quality of Software (S206) project.

The collections include tests for making GET requests to the PokeAPI and asserting the status code and response body. They also include a test for a non-existent Pokemon, which should return a 404 status code.

The collections are useful for automating API testing and ensuring the quality of the software.