feat: Refactor Postman collections for Quality of Software (S206)
This commit refactors the Postman collections for the Quality of Software (S206) project. The changes include removing unnecessary test cases and improving the structure and organization of the collections. This will make it easier to maintain and update the API tests in the future.

The refactored collections still include tests for making GET requests to the PokeAPI and asserting the status code and response body. They also include a test for a non-existent Pokemon, which should return a 404 status code.

These changes improve the quality and maintainability of the software.